Added static getDefaultPos/Scale/Dir/Up to class D3PosDirUpScale Changed default-values:defdaultDir changed from (0,0,-1)->(0,1,0) defaultUp changed from (0,1,0)->(0,0,1) Changed most double-values in D3Math to float Changed D3Spherical parent-class from Spherical->FloatSpherical Changed SphericalTemplate-constructors to be function-templates Added field BYTE m_stateFlags to both D3Scene and D3SceneEditor, and control all boolean fields by set-,clrFlags,isSet-functions enum StateFlags in D3SceneEdtior.h now macroes 0x01,0x02.... and included SE_PROP_CHANGES, SE_LIGHTCONTROLS,SE__RENDER_3D,SE_RENDER_INFO in same sequence (mssked together in D3SceneEditor::m_stateFlags) Same construction with SC_INITDONE,SC_PROPCHANGES,SC_RIGHTHANDED... masked in D3Scene::m_stateFlags Modified container of used materials to typedef CompactUIntHashMap<MATERIAL> MaterialMap and renamed MATERIAL::m_index to m_id. added new functiona to class MATERIAL, (static D3DMATERIAL createMaterialWithColor, Added typedef CompactArray<LIGHT> LightArray Moved function createDefaultLight from D3Scene->Light.h Added functions D3Scene::removeAllLights, setLightControlMaterial Added typedef LightArray Renamed all camera-controlling functions to *Camera* to *Cam* Added file D3Ray.cpp containing all functionality about picking objects with mouse Added function D3SceneEditor::checkInvariant which is called by MACRO CHECKINVARIANT in Debug-mode Removed member m_materialIndex from D3SceneObjectLineArrow Refactored D3SceneObjectCoordinateSystem

+++ Lib/Src/D3DGraphics/D3AnimatedSurface.cpp
@@ -37,7 +37,7 @@ void D3AnimatedSurface::stopAnimation() {
   }
 }
 
-void D3AnimatedSurface::scaleSpeed(double factor) {
+void D3AnimatedSurface::scaleSpeed(float factor) {
   if((factor == 0) || (factor == 1)) {
     return;
   }
